Move bqrs-utils.ts to /common
