+++ extensions/ql-vscode/src/language-support/contextual/file-range-from-uri.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 
 import { UrlValue, LineColumnLocation } from "../../common/bqrs-cli-types";
-import { isEmptyPath } from "../../pure/bqrs-utils";
+import { isEmptyPath } from "../../common/bqrs-utils";
 import { DatabaseItem } from "../../databases/local-databases";
 
 export function fileRangeFromURI(
