+++ extensions/ql-vscode/src/view/results/graph.tsx
@@ -6,7 +6,7 @@ import {
   GraphInterpretationData,
 } from "../../pure/interface-types";
 import { graphviz, GraphvizOptions } from "d3-graphviz";
-import { tryGetLocationFromString } from "../../pure/bqrs-utils";
+import { tryGetLocationFromString } from "../../common/bqrs-utils";
 export type GraphProps = ResultTableProps & {
   resultSet: InterpretedResultSet<GraphInterpretationData>;
 };
