+++ extensions/ql-vscode/src/view/results/raw-results-table.tsx
@@ -16,7 +16,7 @@ import RawTableHeader from "./RawTableHeader";
 import RawTableRow from "./RawTableRow";
 import { ResultRow } from "../../common/bqrs-cli-types";
 import { onNavigation } from "./results";
-import { tryGetResolvableLocation } from "../../pure/bqrs-utils";
+import { tryGetResolvableLocation } from "../../common/bqrs-utils";
 import { ScrollIntoViewHelper } from "./scroll-into-view-helper";
 import { sendTelemetry } from "../common/telemetry";
