+++ extensions/ql-vscode/src/view/variant-analysis/RawResultsTable.tsx
@@ -7,7 +7,7 @@ import {
   RawResultSet,
   ResultSetSchema,
 } from "../../common/bqrs-cli-types";
-import { tryGetRemoteLocation } from "../../pure/bqrs-utils";
+import { tryGetRemoteLocation } from "../../common/bqrs-utils";
 import TextButton from "../common/TextButton";
 import { convertNonPrintableChars } from "../../pure/text-utils";
 import { sendTelemetry, useTelemetryOnChange } from "../common/telemetry";
